postgreSQL examples

+++ mybatis-r2dbc-example/src/main/java/pro/chenggang/example/reactive/mybatis/r2dbc/entity/model/Dept.java
@@ -26,7 +26,7 @@ public class Dept {
     /**
      * dept no
      */
-    protected Long deptNo;
+    protected Integer deptNo;
 
     /**
      * dept name
