+++ mybatis-r2dbc-example/src/main/java/pro/chenggang/example/reactive/mybatis/r2dbc/mapper/dynamic/EmpDynamicSqlSupport.java
@@ -10,7 +10,7 @@ import org.mybatis.dynamic.sql.SqlTable;
 public final class EmpDynamicSqlSupport {
     public static final Emp emp = new Emp();
 
-    public static final SqlColumn<Long> empNo = emp.empNo;
+    public static final SqlColumn<Integer> empNo = emp.empNo;
 
     public static final SqlColumn<String> empName = emp.empName;
 
@@ -24,12 +24,12 @@ public final class EmpDynamicSqlSupport {
 
     public static final SqlColumn<BigDecimal> kpi = emp.kpi;
 
-    public static final SqlColumn<Long> deptNo = emp.deptNo;
+    public static final SqlColumn<Integer> deptNo = emp.deptNo;
 
     public static final SqlColumn<LocalDateTime> createTime = emp.createTime;
 
     public static final class Emp extends SqlTable {
-        public final SqlColumn<Long> empNo = column("emp_no", JDBCType.BIGINT);
+        public final SqlColumn<Integer> empNo = column("emp_no", JDBCType.INTEGER);
 
         public final SqlColumn<String> empName = column("emp_name", JDBCType.VARCHAR);
 
@@ -41,9 +41,9 @@ public final class EmpDynamicSqlSupport {
 
         public final SqlColumn<Integer> salary = column("salary", JDBCType.INTEGER);
 
-        public final SqlColumn<BigDecimal> kpi = column("kpi", JDBCType.DECIMAL);
+        public final SqlColumn<BigDecimal> kpi = column("kpi", JDBCType.NUMERIC);
 
-        public final SqlColumn<Long> deptNo = column("dept_no", JDBCType.BIGINT);
+        public final SqlColumn<Integer> deptNo = column("dept_no", JDBCType.INTEGER);
 
         public final SqlColumn<LocalDateTime> createTime = column("create_time", JDBCType.TIMESTAMP);
