+++ mybatis-r2dbc-spring-example/src/main/java/pro/chenggang/example/reactive/mybatis/r2dbc/spring/mapper/dynamic/DeptDynamicSqlSupport.java
@@ -1,15 +1,14 @@
 package pro.chenggang.example.reactive.mybatis.r2dbc.spring.mapper.dynamic;
 
-import org.mybatis.dynamic.sql.SqlColumn;
-import org.mybatis.dynamic.sql.SqlTable;
-
 import java.sql.JDBCType;
 import java.time.LocalDateTime;
+import org.mybatis.dynamic.sql.SqlColumn;
+import org.mybatis.dynamic.sql.SqlTable;
 
 public final class DeptDynamicSqlSupport {
     public static final Dept dept = new Dept();
 
-    public static final SqlColumn<Long> deptNo = dept.deptNo;
+    public static final SqlColumn<Integer> deptNo = dept.deptNo;
 
     public static final SqlColumn<String> deptName = dept.deptName;
 
@@ -18,7 +17,7 @@ public final class DeptDynamicSqlSupport {
     public static final SqlColumn<LocalDateTime> createTime = dept.createTime;
 
     public static final class Dept extends SqlTable {
-        public final SqlColumn<Long> deptNo = column("dept_no", JDBCType.BIGINT);
+        public final SqlColumn<Integer> deptNo = column("dept_no", JDBCType.INTEGER);
 
         public final SqlColumn<String> deptName = column("dept_name", JDBCType.VARCHAR);
